fix(MOR-242): arm USB-audio PCM TX in shared serial Icom base
start_audio_tx_pcm in _IcomSerialRadioBase opened the USB CODEC TX stream
but never set _pcm_tx_fmt, and the base had no push_audio_tx_pcm override,
so pushes fell through to AudioRuntimeMixin.push_audio_tx_pcm whose guard
raised "PCM TX not started" on every frame. This affected every non-7610
serial Icom backend (IC-705/X6200, IC-7300, IC-9700); the working path
lived only in Icom7610SerialRadio.

Lift the working TX path into the shared base:
- start_audio_tx_pcm clamps channels to 1 (USB CODEC mic is mono-only;
  GH#1382, consistent with the MOR-238 RX clamp) and arms _pcm_tx_fmt.
- add push_audio_tx_pcm that validates frame size against the armed
  format and forwards to the USB audio driver.
- add start/push/stop_audio_tx_opus.
- clear _pcm_tx_fmt in stop_audio_tx_pcm and _stop_serial_audio_driver.

Delete the now-duplicate overrides from Icom7610SerialRadio so it inherits
the base; behavior is pinned identical by the unchanged
tests/test_icom7610_serial_radio.py.
